refactor: move uuid logic to src/utils/uuid.ts
Amp-Thread-ID: https://ampcode.com/threads/T-019e70e4-c5ae-715c-8800-4d0f23c789e4
